+++ sdk/resourcemanager/azure-resourcemanager-compute/CHANGELOG.md
@@ -1,16 +1,26 @@
 # Release History
 
-## 2.11.0-beta.1 (Unreleased)
+## 2.12.0-beta.1 (Unreleased)
 
 ### Features Added
-
-- Supported Flexible orchestration mode for `VirtualMachineScaleSet` during create.
 
 ### Breaking Changes
 
 ### Bugs Fixed
 
 ### Other Changes
+
+## 2.11.0 (2022-01-17)
+
+### Features Added
+
+- Supported Flexible orchestration mode for `VirtualMachineScaleSet` during create.
+
+### Other Changes
+
+#### Dependency Updates
+
+- Updated `api-version` to `2021-08-01`.
 
 ## 2.10.0 (2021-11-22)