+++ sdk/schemaregistry/azure-data-schemaregistry-apacheavro/CHANGELOG.md
@@ -1,12 +1,16 @@
 # Release History
 
-## 1.0.0-beta.8 (Unreleased)
+## 1.0.0-beta.9 (Unreleased)
 
 ### Features Added
 
 ### Breaking Changes
 
 ### Bugs Fixed
+
+### Other Changes
+
+## 1.0.0-beta.8 (2022-01-18)
 
 ### Other Changes